Add word 'YouTube'
Imagine being in a dark, quiet room, at night, browsing GitHub.
You see a link that you want to click, and unbeknownst to you, it's a link to a YouTube video.
Suddenly, a loud train whistle goes off! "HELLO AND WELCOME TO A CODING CHALLENGE!!"
Your frightened, and immediately close the tab, hoping that no one else heard.

I wouldn't want that to happen to you, so I added the word, YouTube.